Refactor: libpacemaker: separate action-related code into own source file
This moves update_action() and some static functions it uses from
pcmk_graph_producer.c to a new pcmk_sched_actions.c, for more logical
organization (these are really scheduling functions, not directly related to
transition graphs). There are no code changes.

Before:
 1558 lib/pacemaker/pcmk_graph_producer.c

After:
  988 lib/pacemaker/pcmk_graph_producer.c
  599 lib/pacemaker/pcmk_sched_actions.c
